Update project summary service.

+++ Estimation.Domain/Models/ProjectSummary.cs
@@ -12,46 +12,46 @@ namespace Estimation.Domain.Models
         /// <summary>
         /// Sum of material price
         /// </summary>
-        public decimal MaterialPrice { get; set; }
+        public int MaterialPrice { get; set; }
 
         /// <summary>
         /// Sum of accessories price
         /// </summary>
-        public decimal Accessories { get; set; }
+        public int Accessories { get; set; }
 
         /// <summary>
         /// Sum of fittings price
         /// </summary>
-        public decimal Fittings { get; set; }
+        public int Fittings { get; set; }
 
         /// <summary>
         /// Sum of supporting price
         /// </summary>
-        public decimal Supporting { get; set; }
+        public int Supporting { get; set; }
 
         /// <summary>
         /// Sum of painting price
         /// </summary>
-        public decimal Painting { get; set; }
+        public int Painting { get; set; }
 
         /// <summary>
         /// Miscellaneous price
         /// </summary>
-        public decimal Miscellaneous { get; set; }
+        public int Miscellaneous { get; set; }
 
         /// <summary>
         /// Sum of installation price
         /// </summary>
-        public decimal Installation { get; set; }
+        public int Installation { get; set; }
 
         /// <summary>
         /// Transportation price
         /// </summary>
-        public decimal Transportation { get; set; }
+        public int Transportation { get; set; }
 
         /// <summary>
         /// Overall price
         /// </summary>
-        public decimal GrandTotal { get; set; }
+        public int GrandTotal { get; set; }
     }
 }
